Move suite types and instances into respective version modules
