+++ rustls/src/client/hs.rs
@@ -4,7 +4,6 @@ use crate::check::check_message;
 use crate::conn::{ConnectionCommon, ConnectionRandoms};
 use crate::error::{Error, WebPkiError};
 use crate::hash_hs::HandshakeHashBuffer;
-use crate::tls13::key_schedule::KeyScheduleEarly;
 use crate::kx;
 #[cfg(feature = "logging")]
 use crate::log::{debug, trace};
@@ -25,6 +24,7 @@ use crate::msgs::handshake::{Random, SessionID};
 use crate::msgs::message::{Message, MessagePayload};
 use crate::msgs::persist;
 use crate::ticketer::TimeBase;
+use crate::tls13::key_schedule::KeyScheduleEarly;
 use crate::SupportedCipherSuite;
 
 #[cfg(feature = "tls12")]
