+++ rustls/src/lib.rs
@@ -328,12 +328,13 @@ pub use crate::server::StoresServerSessions;
 pub use crate::server::{ClientHello, ProducesTickets, ResolvesServerCert};
 pub use crate::server::{ServerConfig, ServerConnection};
 pub use crate::stream::{Stream, StreamOwned};
-#[cfg(feature = "tls12")]
-pub use crate::suites::Tls12CipherSuite;
 pub use crate::suites::{
-    BulkAlgorithm, SupportedCipherSuite, Tls13CipherSuite, ALL_CIPHER_SUITES, DEFAULT_CIPHER_SUITES,
+    BulkAlgorithm, SupportedCipherSuite, ALL_CIPHER_SUITES, DEFAULT_CIPHER_SUITES,
 };
 pub use crate::ticketer::Ticketer;
+#[cfg(feature = "tls12")]
+pub use crate::tls12::Tls12CipherSuite;
+pub use crate::tls13::Tls13CipherSuite;
 pub use crate::verify::{
     AllowAnyAnonymousOrAuthenticatedClient, AllowAnyAuthenticatedClient, NoClientAuth,
 };
@@ -343,21 +344,21 @@ pub use crate::versions::{SupportedProtocolVersion, ALL_VERSIONS, DEFAULT_VERSIO
 ///
 /// [`ALL_CIPHER_SUITES`] is provided as an array of all of these values.
 pub mod cipher_suite {
-    pub use crate::suites::TLS13_AES_128_GCM_SHA256;
-    pub use crate::suites::TLS13_AES_256_GCM_SHA384;
-    pub use crate::suites::TLS13_CHACHA20_POLY1305_SHA256;
     #[cfg(feature = "tls12")]
-    pub use crate::suites::TLS_ECDHE_ECDSA_WITH_AES_128_GCM_SHA256;
+    pub use crate::tls12::TLS_ECDHE_ECDSA_WITH_AES_128_GCM_SHA256;
     #[cfg(feature = "tls12")]
-    pub use crate::suites::TLS_ECDHE_ECDSA_WITH_AES_256_GCM_SHA384;
+    pub use crate::tls12::TLS_ECDHE_ECDSA_WITH_AES_256_GCM_SHA384;
     #[cfg(feature = "tls12")]
-    pub use crate::suites::TLS_ECDHE_ECDSA_WITH_CHACHA20_POLY1305_SHA256;
+    pub use crate::tls12::TLS_ECDHE_ECDSA_WITH_CHACHA20_POLY1305_SHA256;
     #[cfg(feature = "tls12")]
-    pub use crate::suites::TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256;
+    pub use crate::tls12::TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256;
     #[cfg(feature = "tls12")]
-    pub use crate::suites::TLS_ECDHE_RSA_WITH_AES_256_GCM_SHA384;
+    pub use crate::tls12::TLS_ECDHE_RSA_WITH_AES_256_GCM_SHA384;
     #[cfg(feature = "tls12")]
-    pub use crate::suites::TLS_ECDHE_RSA_WITH_CHACHA20_POLY1305_SHA256;
+    pub use crate::tls12::TLS_ECDHE_RSA_WITH_CHACHA20_POLY1305_SHA256;
+    pub use crate::tls13::TLS13_AES_128_GCM_SHA256;
+    pub use crate::tls13::TLS13_AES_256_GCM_SHA384;
+    pub use crate::tls13::TLS13_CHACHA20_POLY1305_SHA256;
 }
 
 /// All defined protocol versions appear in this module.
